requirements.txt
aliyun-python-sdk-core==2.13.13
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.8.0
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.5
django-debug-toolbar==2.1
django-filter==2.2.0
django-haystack==2.8.1
django-import-export==2.0
django-redis==4.11.0
django-simpleui==3.8
django-tinymce==2.8.0
djangorestframework==3.10.3
djangorestframework-stubs==1.1.0
drf-haystack==1.8.6
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
jieba==0.39
Jinja2==2.10.3
jmespath==0.9.4
kombu==4.6.7
MarkupPy==1.14
MarkupSafe==1.1.1
more-itertools==8.0.2
mutagen==1.43.0
odfpy==1.4.0
openpyxl==3.0.2
oss2==2.9.0
Pillow==6.2.1
pycparser==2.19
pycryptodome==3.9.4
pycryptodomex==3.7.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==1.10.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
